+++ mitiq/pec/pec.py
@@ -15,7 +15,7 @@
 
 """High-level probabilistic error cancellation tools."""
 
-from typing import Optional, Callable, List, Union, Tuple
+from typing import Optional, Callable, List, Union, Tuple, Dict, Any
 import warnings
 
 import numpy as np
@@ -46,8 +46,8 @@ def execute_with_pec(
     force_run_all: bool = True,
     random_state: Optional[Union[int, np.random.RandomState]] = None,
     full_output: bool = False,
-) -> Union[float, Tuple[float, float]]:
-    """Evaluates the expectation value associated to the input circuit
+) -> Union[float, Tuple[float, Dict[str, Any]]]:
+    r"""Evaluates the expectation value associated to the input circuit
     using probabilistic error cancellation (PEC) [Temme2017]_ [Endo2018]_.
 
     This function implements PEC by:
@@ -77,17 +77,17 @@ def execute_with_pec(
             uniqueness, else a minimal unique set is executed.
         random_state: Seed for sampling circuits.
         full_output: If False only the average PEC value is returned.
-            If True an estimate of the associated error is returned too.
+            If True a dictionary containing all PEC data is returned too.
 
     Returns:
         pec_value: The PEC estimate of the ideal expectation value associated
             to the input circuit.
-        pec_error: The estimated error between the mitigated 'pec_value' and
-            the actual ideal expectation value. This is estimated as the ratio
-            pec_std / sqrt(num_samples), where 'pec_std' is the
-            standard deviation of the PEC samples, i.e., the square root of
-            the mean squared deviation of the sampled values from 'pec_value'.
-            This is returned only if 'full_output' is True.
+        pec_data: A dictionary which contains all the raw data involved in the
+            PEC process (including the PEC estimation error). The error is
+            estimated as pec_std / sqrt(num_samples), where 'pec_std' is the
+            standard deviation of the PEC samples, i.e., the square root of the
+            mean squared deviation of the sampled values from 'pec_value'.
+            This is returned only if ``full_output`` is ``True``.
 
     .. [Temme2017] : Kristan Temme, Sergey Bravyi, Jay M. Gambetta,
         "Error mitigation for short-depth quantum circuits,"
@@ -144,8 +144,20 @@ def execute_with_pec(
 
     pec_value = np.average(unbiased_estimators)
 
-    if full_output:
-        pec_error = np.std(unbiased_estimators) / np.sqrt(num_samples)
-        return pec_value, pec_error
+    if not full_output:
+        return pec_value
 
-    return pec_value
+    # Build dictionary with additional results and data
+    pec_data: Dict[str, Any] = {}
+
+    pec_data = {
+        "num_samples": num_samples,
+        "precision": precision,
+        "pec_value": pec_value,
+        "pec_error": np.std(unbiased_estimators) / np.sqrt(num_samples),
+        "unbiased_estimators": unbiased_estimators,
+        "measured_expectation_values": exp_values,
+        "sampled_circuits": sampled_circuits,
+    }
+
+    return pec_value, pec_data